+++ minitrace-macro/tests/ui/ok/has-name-and-full-name.rs
@@ -1,6 +1,6 @@
 use minitrace::trace;
 
-#[trace(path_name = true)]
+#[trace(name = "Name", full_name = false)]
 async fn f(a: u32) -> u32 {
     a
 }